% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.diversity.R
\name{chao1_alpha}
\alias{chao1_alpha}
\title{Calculate Chao1 Richness Estimator}
\usage{
chao1_alpha(x)
}
\arguments{
\item{x}{A numeric vector of counts.}
}
\value{
A numeric value representing the Chao1 estimate of richness.
}
\description{
Computes the Chao1 estimator for species richness based on observed singleton and doubleton species.
}
\examples{
chao1_alpha(c(1, 1, 2, 3, 4))
}
