% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.diversity.R
\name{chao1_bc_alpha}
\alias{chao1_bc_alpha}
\title{Calculate Bias-Corrected Chao1 Richness Estimator}
\usage{
chao1_bc_alpha(x)
}
\arguments{
\item{x}{A numeric vector of counts.}
}
\value{
A numeric value representing the bias-corrected Chao1 estimate.
}
\description{
Computes a bias-corrected version of the Chao1 richness estimator to account for small sample sizes.
}
\examples{
chao1_bc_alpha(c(1, 1, 2, 3, 4))
}
