% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.diversity.R
\name{gini_simpson_alpha}
\alias{gini_simpson_alpha}
\title{Calculate Gini-Simpson Diversity Index}
\usage{
gini_simpson_alpha(x)
}
\arguments{
\item{x}{A numeric vector of counts.}
}
\value{
A numeric value representing the Gini-Simpson index.
}
\description{
Computes the Gini-Simpson index, a measure of diversity representing the probability that two randomly selected individuals belong to different species.
}
\examples{
gini_simpson_alpha(c(5, 10, 15))
}
