% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.diversity.R
\name{pielou_alpha}
\alias{pielou_alpha}
\title{Calculate Pielou's Evenness Index}
\usage{
pielou_alpha(x)
}
\arguments{
\item{x}{A numeric vector of counts.}
}
\value{
A numeric value representing Pielou's evenness index.
}
\description{
Computes Pielou's evenness, which quantifies how evenly individuals are distributed across species.
}
\examples{
pielou_alpha(c(10, 10, 10))
}
