% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.diversity.R
\name{unique_seq_alpha}
\alias{unique_seq_alpha}
\title{Calculate Number of Unique Sequences}
\usage{
unique_seq_alpha(x)
}
\arguments{
\item{x}{A numeric vector of counts (e.g., sequence or OTU abundances).}
}
\value{
An integer indicating the number of unique (nonzero) sequences.
}
\description{
Computes the number of observed unique sequences, i.e., how many elements
in the count vector have nonzero abundance.
}
\examples{
unique_seq_alpha(c(1, 0, 2, 3, 0))
# Returns 3

unique_seq_alpha(c(0, 0, 0))
# Returns 0

}
